hopefully fixed linux builds...

+++ src/utils/utils_tests.cpp
@@ -66,20 +66,20 @@ test_case("[utils] numDigits") {
     check_eq(numDigits(235576746), 9u);
     check_eq(numDigits(2125767446), 10u);
 
-    check_eq(numDigits(21257674464LL), 11u);
-    check_eq(numDigits(212576744643LL), 12u);
-    check_eq(numDigits(2125767446454LL), 13u);
-    check_eq(numDigits(21257674456564LL), 14u);
-    check_eq(numDigits(212534567674464LL), 15u);
-    check_eq(numDigits(2125767546354464LL), 16u);
-    check_eq(numDigits(21257674445656664LL), 17u);
-    check_eq(numDigits(212576744656566664LL), 18u);
-    check_eq(numDigits(9223372036854775807LL), 19u);
+    check_eq(numDigits(int64(21257674464LL)), 11u);
+    check_eq(numDigits(int64(212576744643LL)), 12u);
+    check_eq(numDigits(int64(2125767446454LL)), 13u);
+    check_eq(numDigits(int64(21257674456564LL)), 14u);
+    check_eq(numDigits(int64(212534567674464LL)), 15u);
+    check_eq(numDigits(int64(2125767546354464LL)), 16u);
+    check_eq(numDigits(int64(21257674445656664LL)), 17u);
+    check_eq(numDigits(int64(212576744656566664LL)), 18u);
+    check_eq(numDigits(int64(9223372036854775807LL)), 19u);
 }
 
 test_case("[utils] itoa_fast") {
-#define itoa_fast_check_eq(x) check_eq(strcmp(itoa_fast(HA_CAT_1(x, LL), dest), #x), 0)
-#define itoa_fast_check_ne(x) check_ne(strcmp(itoa_fast(HA_CAT_1(x, LL), dest), #x), 0)
+#define itoa_fast_check_eq(x) check_eq(strcmp(itoa_fast(int64(HA_CAT_1(x, LL)), dest), #x), 0)
+#define itoa_fast_check_ne(x) check_ne(strcmp(itoa_fast(int64(HA_CAT_1(x, LL)), dest), #x), 0)
     char dest[24];
     ((void)dest); // for the unused variable warning when tests are disabled
     itoa_fast_check_eq(0);
